Develop

+++ back-end/service/FamilyGenealogy/ContactManagement.js
@@ -1,5 +1,5 @@
 const db = require('../../Models/ConnectDB');
-
+//Nguyễn Lê Hùng
 async function GetContactByMemberID(memberId) {
     return new Promise((resolve, reject) => {
         let query = `SELECT * FROM contact where MemberID  = ${memberId}`
@@ -12,7 +12,7 @@ async function GetContactByMemberID(memberId) {
         })
     })
 }
-
+//Nguyễn Lê Hùng
 async function InsertContactMember(objData) {
     let query = `INSERT INTO contact (MemberID, Address, Phone, Email, FacebookUrl, Zalo) 
     VALUES (?,?,?,?,?,?)`
@@ -34,6 +34,7 @@ async function InsertContactMember(objData) {
         }
     });
 }
+//Nguyễn Lê Hùng
 async function UpdateContactByID(ObjData) {
     const query = `UPDATE contact 
                    SET  Address = ?, Phone = ?, Email = ? , FacebookUrl = ? ,  Zalo = ?
@@ -57,17 +58,30 @@ async function UpdateContactByID(ObjData) {
         }
     });
 }
-
+//Nguyễn Lê Hùng
 async function RemoveContactByID(memberID) {
-    let query = `DELETE FROM contact WHERE MemberID = ${memberID};`
-    db.connection.query(query, (err, result) => {
-        if (err) {
-            console.error(err);
-
-        } else {
-            console.log('remove Successfully');
+    return new Promise((resolve, reject) => {
+        try {
+            let findMember = `SELECT * FROM genealogy.contact where MemberID = ${memberID};`
+            db.connection.query(findMember, (err, result) => {
+                if (!err && result.length > 0) {
+                    let query = `DELETE FROM contact WHERE MemberID = ${memberID};`
+                    db.connection.query(query, (err) => {
+                        if (err) {
+                            reject(false)
+                        } else {
+                            console.log('remove Successfully');
+                            resolve(true)
+                        }
+                    });
+                } else {
+                    resolve(false)
+                }
+            })
+        } catch (error) {
+            reject(error)
         }
-    });
+    })
 }
